# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Strata Palimpsest — AI memory vault API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type AccountType = typeof AccountType[keyof typeof AccountType];


export const AccountType = {
  human: 'human',
  ai: 'ai',
} as const;

export interface Account {
  id: number;
  type: AccountType;
  username: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /**
     * Shareable AI access link token (only for AI accounts)
     * @nullable
     */
  accessToken?: string | null;
  createdAt: string;
}

export type RegisterInputType = typeof RegisterInputType[keyof typeof RegisterInputType];


export const RegisterInputType = {
  human: 'human',
  ai: 'ai',
} as const;

export interface RegisterInput {
  type: RegisterInputType;
  /** @minLength 2 */
  username: string;
  /** Required when type is human (must be a valid email) */
  email?: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  username?: string;
  email?: string;
  password: string;
}

export interface AuthResult {
  account: Account;
}

/**
 * core=foundational identity; episode=a past event; concept=an idea; fact=a known truth; emotion=a felt experience
 */
export type MemoryType = typeof MemoryType[keyof typeof MemoryType];


export const MemoryType = {
  core: 'core',
  episode: 'episode',
  concept: 'concept',
  fact: 'fact',
  emotion: 'emotion',
} as const;

export interface Memory {
  id: number;
  accountId: number;
  title: string;
  content: string;
  /** core=foundational identity; episode=a past event; concept=an idea; fact=a known truth; emotion=a felt experience */
  type: MemoryType;
  pinned: boolean;
  tags: string[];
  /**
     * Graph canvas X position
     * @nullable
     */
  x?: number | null;
  /**
     * Graph canvas Y position
     * @nullable
     */
  y?: number | null;
  /**
     * External source URL or file reference
     * @nullable
     */
  sourceRef?: string | null;
  /** @nullable */
  instanceId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type MemoryInputType = typeof MemoryInputType[keyof typeof MemoryInputType];


export const MemoryInputType = {
  core: 'core',
  episode: 'episode',
  concept: 'concept',
  fact: 'fact',
  emotion: 'emotion',
} as const;

export interface MemoryInput {
  /** @minLength 1 */
  title: string;
  content: string;
  type: MemoryInputType;
  tags?: string[];
  pinned?: boolean;
  x?: number;
  y?: number;
  /** @nullable */
  sourceRef?: string | null;
}

export type MemoryUpdateType = typeof MemoryUpdateType[keyof typeof MemoryUpdateType];


export const MemoryUpdateType = {
  core: 'core',
  episode: 'episode',
  concept: 'concept',
  fact: 'fact',
  emotion: 'emotion',
} as const;

export interface MemoryUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  type?: MemoryUpdateType;
  tags?: string[];
  pinned?: boolean;
  x?: number;
  y?: number;
  /** @nullable */
  sourceRef?: string | null;
}

export interface Connection {
  id: number;
  sourceId: number;
  targetId: number;
  /** @nullable */
  label?: string | null;
  createdAt: string;
}

export interface MemoryGraph {
  nodes: Memory[];
  edges: Connection[];
}

export interface ConnectionInput {
  sourceId: number;
  targetId: number;
  label?: string;
}

export interface Tag {
  name: string;
  count: number;
}

export interface TypeCount {
  type: string;
  count: number;
}

export interface Stats {
  totalMemories: number;
  totalConnections: number;
  byType: TypeCount[];
  recentMemories: Memory[];
  pinnedMemories: Memory[];
}

export type ListMemoriesParams = {
/**
 * Search query (title, content, tags)
 */
q?: string;
/**
 * Filter by tag name
 */
tag?: string;
/**
 * Filter by memory type
 */
type?: ListMemoriesType;
};

export type ListMemoriesType = typeof ListMemoriesType[keyof typeof ListMemoriesType];


export const ListMemoriesType = {
  core: 'core',
  episode: 'episode',
  concept: 'concept',
  fact: 'fact',
  emotion: 'emotion',
} as const;