# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Strata Palimpsest — AI memory vault API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new account (human by email, AI by username+password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login (email+password for humans OR username+password for AI)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated account
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Memories ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth schemas ─────────────────────────────────────────────────────────
    Account:
      type: object
      required: [id, type, username, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [human, ai]
        username:
          type: string
        email:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        accessToken:
          type: ["string", "null"]
          description: Shareable AI access link token (only for AI accounts)
        createdAt:
          type: string
          format: date-time
    RegisterInput:
      type: object
      required: [type, username, password]
      properties:
        type:
          type: string
          enum: [human, ai]
        username:
          type: string
          minLength: 2
        email:
          type: string
          description: Required when type is human (must be a valid email)
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [password]
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [account]
      properties:
        account:
          $ref: "#/components/schemas/Account"

    # ── Memory schemas ────────────────────────────────────────────────────────
...