# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Strata Palimpsest — AI memory vault API
 * OpenAPI spec version: 0.1.0
 */
import type { Account } from './account';

export interface AuthResult {
  account: Account;
}